+++ internal/resolvers/default/resolver.go
@@ -9,9 +9,11 @@ import (
 	"github.com/Chatterino/api/internal/resolvers/betterttv"
 	"github.com/Chatterino/api/internal/resolvers/discord"
 	"github.com/Chatterino/api/internal/resolvers/frankerfacez"
+	"github.com/Chatterino/api/internal/resolvers/imgur"
 	"github.com/Chatterino/api/internal/resolvers/supinic"
 	"github.com/Chatterino/api/internal/resolvers/twitch"
 	"github.com/Chatterino/api/internal/resolvers/twitter"
+	"github.com/Chatterino/api/internal/resolvers/wikipedia"
 	"github.com/Chatterino/api/internal/resolvers/youtube"
 	"github.com/Chatterino/api/pkg/cache"
 	"github.com/Chatterino/api/pkg/resolver"
@@ -82,6 +84,8 @@ func New(baseURL string) *R {
 	r.customResolvers = append(r.customResolvers, youtube.New()...)
 	r.customResolvers = append(r.customResolvers, supinic.New()...)
 	r.customResolvers = append(r.customResolvers, twitch.New()...)
+	r.customResolvers = append(r.customResolvers, imgur.New()...)
+	r.customResolvers = append(r.customResolvers, wikipedia.New()...)
 
 	return r
 }