+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Unreleased
 
+- Added support for Wikipedia article links. (#92)
+- Fixed an issue where OpenGraph descriptions were not HTML-sanitized. (#90)
+- Added support for imgur image links. (#81)
 - Added support for FrankerFaceZ emote links. (#57)
 - Added author name to Twitch clips response. (#76)